fix size_t - int issue again

+++ testing/adios2/engine/bp/TestBPStepsInSituLocalArray.cpp
@@ -160,8 +160,8 @@ TEST_P(BPStepsInSituLocalArrayReaders, EveryStep)
     adios2::Engine reader = ior.Open(fname, adios2::Mode::Read);
     EXPECT_TRUE(reader);
 
-    size_t stepsWritten = 0;
-    size_t stepsRead = 0;
+    int stepsWritten = 0;
+    int stepsRead = 0;
 
     for (const auto act : schedule)
     {
@@ -288,8 +288,8 @@ TEST_P(BPStepsInSituLocalArrayReaders, NewVarPerStep)
     adios2::Engine reader = ior.Open(fname, adios2::Mode::Read);
     EXPECT_TRUE(reader);
 
-    size_t stepsWritten = 0;
-    size_t stepsRead = 0;
+    int stepsWritten = 0;
+    int stepsRead = 0;
 
     for (const auto act : schedule)
     {
@@ -421,7 +421,7 @@ TEST_P(BPStepsInSituLocalArrayParameters, EveryOtherStep)
     EXPECT_TRUE(writer);
     auto var_i32 = iow.DefineVariable<int32_t>("i32", shape, start, count);
     EXPECT_TRUE(var_i32);
-    auto var_step = iow.DefineVariable<size_t>("step");
+    auto var_step = iow.DefineVariable<int>("step");
     EXPECT_TRUE(var_step);
 
 #if ADIOS2_USE_MPI
@@ -438,10 +438,10 @@ TEST_P(BPStepsInSituLocalArrayParameters, EveryOtherStep)
     adios2::Engine reader = ior.Open(fname, adios2::Mode::Read);
     EXPECT_TRUE(reader);
 
-    size_t stepsWritten = 0;
-    size_t stepsRead = 0;
-    size_t varStepsWritten = 0;
-    size_t varStepsRead = 0;
+    int stepsWritten = 0;
+    int stepsRead = 0;
+    int varStepsWritten = 0;
+    int varStepsRead = 0;
     for (const auto act : schedule)
     {
         if (act == Act::Write)
